Added all the contents of the project version 0.1
Yes, this is not even version 1. Just an initial change and the very
skeleton of the intended app.
The dream and the intentions will be shared in some time.

tl;dr: Have a super simple, command based, cloud syncd, responsive web
app for note taking.

Must be able to handle entire app without touching mouse! like a unix
terminal.